accounts/btceth-wallet-gen/python/requirements.txt
base58==0.2.5
ecdsa==0.13
pysha3==1.0.2
